Add files via upload
Object class is present in java.lang package. Every class in Java is directly or indirectly derived from the Object class. If a class does not extend any other class then it is a direct child class of Object and if extends another class then it is indirectly derived. Therefore the Object class methods are available to all Java classes. Hence Object class acts as a root of the inheritance hierarchy in any Java Program.

The Object class provides multiple methods which are as follows:

*tostring() method
*hashCode() method
*equals(Object obj) method
*finalize() method